refactor(transforms): :truck: move registry up a level
