+++ client/src/registry.ts
