+++ client/src/transformer/transforms/vue-property-decorator/Emit.ts
@@ -6,7 +6,7 @@ import {
   getDecorators,
   isStringLit,
 } from "@/helpers/tsHelpers.js";
-import { registerDecorator } from "@/transformer/registry.js";
+import { registerDecorator } from "@/registry.js";
 import { VxReferenceKind, VxResultKind, VxResultToMacro, VxTransform } from "@/types.js";
 import ts, { TupleTypeNode } from "typescript";
 import { instanceDependencies } from "../utils/instancePropertyAccess.js";
