+++ client/src/transformer/transforms/vue-property-decorator/Inject.ts
@@ -6,7 +6,7 @@ import {
   isStringLit,
 } from "@/helpers/tsHelpers.js";
 import { namedImports } from "@/helpers/utils.js";
-import { registerDecorator } from "@/transformer/registry.js";
+import { registerDecorator } from "@/registry.js";
 import { VxReferenceKind, VxResultKind, VxTransform } from "@/types.js";
 import { cloneNode } from "ts-clone-node";
 import ts from "typescript";
