+++ client/src/transformer/transforms/vue-property-decorator/Provide.ts
@@ -1,6 +1,6 @@
 import { createCallExpression, createIdentifier, getDecorators } from "@/helpers/tsHelpers.js";
 import { namedImports } from "@/helpers/utils.js";
-import { registerDecorator } from "@/transformer/registry.js";
+import { registerDecorator } from "@/registry.js";
 import { VxReferenceKind, VxResultKind, VxTransform } from "@/types.js";
 import ts from "typescript";
